Resuming development.
A lot of things were changed in this commit. Maybe a lot of things got
broken in the proccess. Some things were removed (like some datasources)
but don't worry: they'll be back as soon as they're fixed. The new coding
standards were now applied to the entire codebase, please follow them.
Now we're targeting the 0.3 release as soon as possible, check the roadmap
for more info.